// src/Api/models/coupon-dto-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface CouponDtoResponse
 */
export interface CouponDtoResponse {
    /**
     * 
     * @type {number}
     * @memberof CouponDtoResponse
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof CouponDtoResponse
     */
    'code'?: string;
    /**
     * 
     * @type {number}
     * @memberof CouponDtoResponse
     */
    'discountPercentage'?: number;
    /**
     * 
     * @type {string}
     * @memberof CouponDtoResponse
     */
    'validityStartDate'?: string;
    /**
     * 
     * @type {string}
     * @memberof CouponDtoResponse
     */
    'validityEndDate'?: string;
    /**
     * 
     * @type {number}
     * @memberof CouponDtoResponse
     */
    'minimumOrderValue'?: number;
    /**
     * 
     * @type {string}
     * @memberof CouponDtoResponse
     */
    'status'?: string;
}